% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_zchords}
\alias{encode_zchords}
\title{encode_zchords}
\usage{
encode_zchords(x, y, dx = 1, dy = 0.5 * dx, invalids = TRUE)
}
\arguments{
\item{x}{A numerical vector, normally a longitude}

\item{y}{A numerical vector, normally a latitude}

\item{dx}{Resolution on the x-axis, normally in decimal degrees (default 1 degree)}

\item{dy}{Resolution on the y-axis, normally in decimal degrees (default half of the longitude)}

\item{invalids}{A logical (default TRUE), convert values outside mercator ranges to NA.}
}
\value{
A character vector where the grid "coding" are the midpoints separated
by a colon. This vector can be split up by e.g. the tidyr:separate function.
}
\description{
The function takes x and y values and converts them to a grid whose value
represent the midpoints given dx and dy.

The main code is shamelessly stolen from siggi with some tiny bit of
stardust added (if only in the name).
}
