% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TCR_circos.R
\name{plot_TCR_circos}
\alias{plot_TCR_circos}
\title{Make a circos plot of single cells, showing links between TCRs with identical junctions}
\usage{
\code{
plot_TCR_circos(
     tcr_cells, tcr_links,
     ring_colors=NULL,
     link_colors=NULL, link_width=NULL,
     filename=NULL, plottype="pdf", plotdims=c(12,12))}
}
\arguments{
\item{tcr_cells}{data frame containing information at the cell level. Should include identifiers at minimum, plus optional columns for coloring the cells around the circos plot. Cells will be plotted in the order in which they are listed in this data frame.}

\item{tcr_links}{data frame with links to draw between cells. Should include identifiers for the two cells, plus optional columns for color and/or width of the links to be drawn. Often the output of \code{tabulate_shared_TCR_chains}.}

\item{ring_colors}{optional vector with number(s) or name(s) of the columns in \code{tcr_cells} to use for the colors of the ring. Current version can include up to 2 colors: list the inner ring first, outer ring second.}

\item{link_colors}{optional number or name of the column in \code{tcr_links} to use for the colors of the links. Defaults to NULL, which yields black.}

\item{link_width}{optional number or name of the column in \code{tcr_links} to use for scaling the width of the links. Defaults to NULL, which results in all links having the same width.}

\item{filename}{the path to write the plot to. If provided, the function outputs a pdf of the plot, named "{filename}.pdf".}

\item{plottype}{character string, either "pdf" or "png". The type of plot object to output.}

\item{plotdims}{a numeric vector, the size (in inches) of the plotting object. Either the size of the pdf, or the size of the plotting window.}
}
\description{
This function combines a number of steps for generating a circos plot from TCR sequences. The
purpose is to show clonotype sharing among cells within and between samples. Samples can optionally
be colored by grouping variables, and there are multiple options for determining TCR clonality.
}
