% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_BLAST_TCR_align_hits.R
\name{read_BLAST_TCR_align_hits}
\alias{read_BLAST_TCR_align_hits}
\title{Read the alignment output hit table .txt file from a TCR sequence BLAST query}
\usage{
\code{
  read_BLAST_TCR_align_hits(
    filename, identity_threshold=100, length_threshold=100,
    query_seq_in_name=TRUE)}
}
\arguments{
\item{filename}{character string, the path and file name to read.}

\item{identity_threshold}{numeric, the threshold for \% identity match between the query and the subject sequences. Defaults to 100, can be reduced to allow less strict matching.}

\item{length_threshold}{numeric, the threshold for \% of the total length of the query sequence that must be found in the alignment. Defaults to 100, can be reduced to allow less strict matching. Setting to 0 or NULL results in no thresholding.}

\item{query_seq_in_name}{logical, whether the alignment query sequence is included in the fasta query name. Alignment length thresholding requires the query sequence, which is assumed to be at the end of the query name.}
}
\value{
a data frame containing the alignment hits, optionally filtered by identity and length thresholds
}
\description{
Read in the alignment hit table file output by BLAST following a query of TCR sequences.
This function reads and reformats the alignment results for downstream use. It expects the format
output by the BLAST Download > Alignment > Hit Table(text).
}
