% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_shared_TCR_chains.R
\name{tabulate_shared_TCR_chains}
\alias{tabulate_shared_TCR_chains}
\title{Tabulate the shared TCRs chains in pairs of single-cell libraries}
\usage{
tabulate_shared_TCR_chains(
     tcr_chain_matches,
     tcr1_col="tcr1", tcr2_col="tcr2")
}
\arguments{
\item{tcr_chain_matches}{a data frame containing the TCR chain matches. Should contain, at minimum, the identifiers for the two TCRs. Often the output of \code{match_TCR_chains}.}

\item{tcr1_col, tcr2_col}{numbers or names of the columns containing the identifiers}
}
\value{
A data frame containing the two TCR identifiers and the number of chains shared between them.
}
\description{
Aggregate a set of TCR chain matches to generate a table with pairs of libraries and the number
of chains shared between the libraries.
}
