% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb.R
\name{vb}
\alias{vb}
\title{von Bertalanffy function}
\usage{
vb(age, loo = 120, k = 0.15, to = 0)
}
\arguments{
\item{age}{A numeric vector of age}

\item{loo}{A value for L-infinity}

\item{k}{A value for k}

\item{to}{A value for t-zero}
}
\value{
A vector of derived lengths
}
\description{
Calculates length from age given the von Bertalanffy parameters
}
